feat: unify pipeline adapter
