+++ work-creator/README.md
@@ -1,8 +1,8 @@
 * Start KinD
 * `make deploy` in root of kratix
-* `cd work-creator/pipeline/cmd`
-* Start work creator (note `-input-directory` needs to be an absolute path)
-** `go run main.go -identifier=workName -input-directory=${PWD}/work-creator/test/integration/samples`
+* `cd work-creator`
+* Start work creator (note `--input-directory` needs to be an absolute path)
+** `go run pipeline/cmd/main.go work-creator --input-directory=${PWD}/test/integration/samples --promise-name=yourPromise --pipeline-name=yourPipeline`
 * Edit samples to your liking
 * Re-run work creator to update objects
-** `go run main.go -identifier=workName -input-directory=${PWD}/work-creator/test/integration/samples`
+** `go run pipeline/cmd/main.go work-creator --input-directory=${PWD}/test/integration/samples --promise-name=yourPromise --pipeline-name=yourPipeline`
